[SASS-2414] migrate view content tests from integration to unit
