+++ it/controllers/studentLoans/StudentLoansQuestionControllerISpec.scala
@@ -20,8 +20,6 @@ import forms.studentLoans.StudentLoanQuestionForm
 import models.IncomeTaxUserData
 import models.employment._
 import models.mongo.{EmploymentCYAModel, EmploymentDetails, EmploymentUserData}
-import org.jsoup.Jsoup
-import org.jsoup.nodes.Document
 import play.api.http.HeaderNames
 import play.api.http.Status._
 import play.api.mvc.Result
@@ -35,122 +33,11 @@ import scala.concurrent.Future
 
 class StudentLoansQuestionControllerISpec extends IntegrationTest with ViewHelpers with EmploymentDatabaseHelper {
 
+  override val userScenarios: Seq[UserScenario[_, _]] = Seq.empty
   private val employmentId: String = "1234567890-0987654321"
   private val startDate = s"$taxYear-04-01"
 
   private def url(taxYearUnique: Int): String = fullUrl(studentLoansQuestionPage(taxYearUnique, employmentId))
-
-  trait CommonExpectedResults {
-    val title: String
-    val heading: String
-    val caption: String
-    val paragraphText_1: String
-    val paragraphText_2: String
-
-    val checkboxHint: String
-    val checkboxUgl: String
-    val checkboxUglHint: String
-    val checkboxPgl: String
-    val checkboxPglHint: String
-    val checkboxNo: String
-    val buttonText: String
-    val errorEmpty: String
-    val errorAll: String
-  }
-
-  object ExpectedResultsEnglish extends CommonExpectedResults {
-    override val title: String = "Did you repay any student loan?"
-    override val heading: String = "Did you repay any student loan while employed by Whiterun Guards?"
-    override val caption: String = s"Student loans for 6 April ${taxYearEOY - 1} to 5 April $taxYearEOY"
-    override val paragraphText_1: String = "We only need to know about payments your employer deducted from your salary."
-    override val paragraphText_2: String = "The Student Loans Company would have told you. Check your payslips or P60 for student loan deductions."
-    override val checkboxHint: String = "Select all that apply."
-    override val checkboxUgl: String = "Yes, undergraduate repayments"
-    override val checkboxUglHint: String = "This covers courses like undergraduate degrees (BA, BSc), foundation degrees or Certificates of Higher Education (CertHE)."
-    override val checkboxPgl: String = "Yes, postgraduate repayments"
-    override val checkboxPglHint: String = "This covers courses like master’s or doctorates."
-    override val checkboxNo: String = "No"
-    override val buttonText: String = "Save and continue"
-    override val errorEmpty: String = "Select the types of student loan you repaid, or select \"No\""
-    override val errorAll: String = "Select the types of student loan you repaid, or select \"No\""
-  }
-
-  object ExpectedResultsEnglishAgent extends CommonExpectedResults {
-    override val title: String = "Did your client repay any student loan?"
-    override val heading: String = "Did your client repay any student loan while employed by Whiterun Guards?"
-    override val caption: String = s"Student loans for 6 April ${taxYearEOY - 1} to 5 April $taxYearEOY"
-    override val paragraphText_1: String = "We only need to know about payments their employer deducted from their salary."
-    override val paragraphText_2: String = "The Student Loans Company would have told your client. Check your client’s payslips or P60 for student loan deductions."
-    override val checkboxHint: String = "Select all that apply."
-    override val checkboxUgl: String = "Yes, undergraduate repayments"
-    override val checkboxUglHint: String = "This covers courses like undergraduate degrees (BA, BSc), foundation degrees or Certificates of Higher Education (CertHE)."
-    override val checkboxPgl: String = "Yes, postgraduate repayments"
-    override val checkboxPglHint: String = "This covers courses like master’s or doctorates."
-    override val checkboxNo: String = "No"
-    override val buttonText: String = "Save and continue"
-    override val errorEmpty: String = "Select the types of student loan your client repaid, or select \"No\""
-    override val errorAll: String = "Select the types of student loan your client repaid, or select \"No\""
-  }
-
-  object ExpectedResultsWelsh extends CommonExpectedResults {
-    override val title: String = "A wnaethoch ad-dalu unrhyw fenthyciad myfyriwr?"
-    override val heading: String = "A wnaethoch ad-dalu unrhyw fenthyciad myfyriwr tra’ch bod wedi’ch cyflogi gan Whiterun Guards?"
-    override val caption: String = s"Benthyciadau Myfyrwyr ar gyfer 6 Ebrill ${taxYearEOY - 1} i 5 Ebrill $taxYearEOY"
-    override val paragraphText_1: String = "Rydym ond angen gwybod am daliadau y gwnaeth eich cyflogwr eu didynnu o’ch cyflog."
-    override val paragraphText_2: String = "Byddai’r Cwmni Benthyciadau Myfyrwyr wedi rhoi gwybod i chi am hyn. Gwiriwch eich slipiau cyflog neu P60 am ddidyniadau benthyciad myfyrwyr."
-    override val checkboxHint: String = "Dewiswch bob un sy’n berthnasol."
-    override val checkboxUgl: String = "Iawn, ad-daliadau israddedig"
-    override val checkboxUglHint: String = "Mae hyn yn cynnwys cyrsiau fel graddau israddedig (BA, BSc), graddau sylfaen neu Dystysgrifau Addysg Uwch (CertHE)."
-    override val checkboxPgl: String = "Iawn, ad-daliadau ôl-raddedig"
-    override val checkboxPglHint: String = "Mae hyn yn cynnwys cyrsiau fel graddau meistr neu ddoethuriaeth."
-    override val checkboxNo: String = "Na"
-    override val buttonText: String = "Cadw ac yn eich blaen"
-    override val errorEmpty: String = "Dewiswch y mathau o fenthyciad myfyriwr a ad-dalwyd gennych, neu dewiswch \"Na\""
-    override val errorAll: String = "Dewiswch y mathau o fenthyciad myfyriwr a ad-dalwyd gennych, neu dewiswch \"Na\""
-  }
-
-  object ExpectedResultsWelshAgent extends CommonExpectedResults {
-    override val title: String = "A wnaeth eich cleient ad-dalu unrhyw fenthyciad myfyriwr?"
-    override val heading: String = "A wnaeth eich cleient ad-dalu unrhyw fenthyciad myfyriwr tra ei fod wedi’i gyflogi gan Whiterun Guards?"
-    override val caption: String = s"Benthyciadau Myfyrwyr ar gyfer 6 Ebrill ${taxYearEOY - 1} i 5 Ebrill $taxYearEOY"
-    override val paragraphText_1: String = "Rydym ond angen gwybod am daliadau y gwnaeth ei gyflogwr eu didynnu o’i gyflog."
-    override val paragraphText_2: String = "Byddai’r Cwmni Benthyciadau Myfyrwyr wedi rhoi gwybod i’ch cleient am hyn. Gwiriwch slipiau cyflog neu P60 eich cleient am ddidyniadau benthyciadau myfyrwyr."
-    override val checkboxHint: String = "Dewiswch bob un sy’n berthnasol."
-    override val checkboxUgl: String = "Iawn, ad-daliadau israddedig"
-    override val checkboxUglHint: String = "Mae hyn yn cynnwys cyrsiau fel graddau israddedig (BA, BSc), graddau sylfaen neu Dystysgrifau Addysg Uwch (CertHE)."
-    override val checkboxPgl: String = "Iawn, ad-daliadau ôl-raddedig"
-    override val checkboxPglHint: String = "Mae hyn yn cynnwys cyrsiau fel graddau meistr neu ddoethuriaeth."
-    override val checkboxNo: String = "Na"
-    override val buttonText: String = "Cadw ac yn eich blaen"
-    override val errorEmpty: String = "Dewiswch y mathau o fenthyciad myfyriwr a ad-dalwyd gan eich cleient, neu dewiswch \"Na\""
-    override val errorAll: String = "Dewiswch y mathau o fenthyciad myfyriwr a ad-dalwyd gan eich cleient, neu dewiswch \"Na\""
-  }
-
-  object Selectors {
-    val paragraphSelector = "#main-content > div > div > p:nth-child(2)"
-    val paragraphSelector_2 = "#main-content > div > div > p:nth-child(3)"
-    val checkboxHint = "#studentLoans-hint"
-
-    val checkboxUgl = "#studentLoans"
-    val checkboxUglText = "#main-content > div > div > form > div > div.govuk-checkboxes > div:nth-child(1) > label"
-    val checkboxUglHint = "#studentLoans-item-hint"
-
-    val checkboxPgl = "#studentLoans-2"
-    val checkboxPglText = "#main-content > div > div > form > div > div.govuk-checkboxes > div:nth-child(2) > label"
-    val checkboxPglHint = "#studentLoans-2-item-hint"
-
-    val checkboxN0 = "#studentLoans-4"
-    val checkboxN0Text = "#main-content > div > div > form > div > div.govuk-checkboxes > div:nth-child(4) > label"
-
-    val checkboxHref = "#studentLoans"
-  }
-
-  override val userScenarios: Seq[UserScenario[CommonExpectedResults, CommonExpectedResults]] = Seq(
-    UserScenario(isWelsh = false, isAgent = false, ExpectedResultsEnglish),
-    UserScenario(isWelsh = false, isAgent = true, ExpectedResultsEnglishAgent),
-    UserScenario(isWelsh = true, isAgent = false, ExpectedResultsWelsh),
-    UserScenario(isWelsh = true, isAgent = true, ExpectedResultsWelshAgent)
-  )
 
   ".show" should {
 
@@ -168,7 +55,7 @@ class StudentLoansQuestionControllerISpec extends IntegrationTest with ViewHelpe
         status(result) shouldBe SEE_OTHER
         await(result).header.headers("Location") shouldBe appConfig.incomeTaxSubmissionOverviewUrl(taxYearEOY)
       }
-      
+
       "a user is accessing the page in year" in {
         val request = FakeRequest("GET", studentLoansQuestionPage(taxYear, employmentId)).withHeaders(HeaderNames.COOKIE -> playSessionCookies(taxYear))
 
@@ -181,9 +68,8 @@ class StudentLoansQuestionControllerISpec extends IntegrationTest with ViewHelpe
         status(result) shouldBe SEE_OTHER
         await(result).header.headers("Location") shouldBe appConfig.incomeTaxSubmissionOverviewUrl(taxYear)
       }
-      
-      "a user has no cyadata" in {
 
+      "a user has no cyadata" in {
         val result = {
           dropEmploymentDB()
           authoriseIndividual()
@@ -197,95 +83,67 @@ class StudentLoansQuestionControllerISpec extends IntegrationTest with ViewHelpe
       }
     }
 
-    userScenarios.foreach { scenarioData =>
-      import scenarioData.commonExpectedResults._
-      s"render the page for isAgent: ${scenarioData.isAgent} isWelsh: ${scenarioData.isWelsh} content" when {
-        "there is cya data in session with previous studentLoans" which {
-          lazy val result = {
-            dropEmploymentDB()
-            authoriseAgentOrIndividual(scenarioData.isAgent)
-            insertCyaData(EmploymentUserData(
-              sessionId,
-              mtditid,
-              nino,
-              taxYearEOY,
-              employmentId, isPriorSubmission = false, hasPriorBenefits = false, hasPriorStudentLoans = false,
-              EmploymentCYAModel(
-                EmploymentDetails(
-                  employerName = "Whiterun Guards",
-                  employerRef = Some("223/AB12399"),
-                  startDate = Some(startDate),
-                  didYouLeaveQuestion = Some(false),
-                  taxablePayToDate = Some(3000.00),
-                  totalTaxToDate = Some(300.00),
-                  currentDataIsHmrcHeld = false
-                ),
-                studentLoans = Some(StudentLoansCYAModel(
-                  uglDeduction = true, Some(1000.22), pglDeduction = true, Some(3000.22)
-                ))
-              )
+    "there is cya data in session with previous studentLoans" in {
+      lazy val result = {
+        dropEmploymentDB()
+        authoriseAgentOrIndividual(false)
+        insertCyaData(EmploymentUserData(
+          sessionId,
+          mtditid,
+          nino,
+          taxYearEOY,
+          employmentId, isPriorSubmission = false, hasPriorBenefits = false, hasPriorStudentLoans = false,
+          EmploymentCYAModel(
+            EmploymentDetails(
+              employerName = "Whiterun Guards",
+              employerRef = Some("223/AB12399"),
+              startDate = Some(startDate),
+              didYouLeaveQuestion = Some(false),
+              taxablePayToDate = Some(3000.00),
+              totalTaxToDate = Some(300.00),
+              currentDataIsHmrcHeld = false
+            ),
+            studentLoans = Some(StudentLoansCYAModel(
+              uglDeduction = true, Some(1000.22), pglDeduction = true, Some(3000.22)
             ))
-            userDataStub(IncomeTaxUserData(), nino, taxYearEOY)
+          )
+        ))
+        userDataStub(IncomeTaxUserData(), nino, taxYearEOY)
 
-
-            urlGet(url(taxYearEOY), scenarioData.isWelsh, headers = Seq(HeaderNames.COOKIE -> playSessionCookies(taxYearEOY)))
-          }
-
-          implicit val document: () => Document = () => Jsoup.parse(result.body)
-
-          titleCheck(title, scenarioData.isWelsh)
-          h1Check(heading)
-          captionCheck(caption)
-          textOnPageCheck(paragraphText_1, Selectors.paragraphSelector)
-          textOnPageCheck(paragraphText_2, Selectors.paragraphSelector_2)
-          hintTextCheck(checkboxHint, Selectors.checkboxHint)
-          textOnPageCheck(checkboxUgl, Selectors.checkboxUglText)
-          inputFieldValueCheck("studentLoans[]", Selectors.checkboxUgl, "ugl")
-          hintTextCheck(checkboxUglHint, Selectors.checkboxUglHint)
-          textOnPageCheck(checkboxPgl, Selectors.checkboxPglText)
-          inputFieldValueCheck("studentLoans[]", Selectors.checkboxPgl, "pgl")
-          hintTextCheck(checkboxPglHint, Selectors.checkboxPglHint)
-          textOnPageCheck(checkboxNo, Selectors.checkboxN0Text)
-          inputFieldValueCheck("studentLoans[]", Selectors.checkboxN0, "none")
-          buttonCheck(buttonText)
-        }
+        urlGet(url(taxYearEOY), headers = Seq(HeaderNames.COOKIE -> playSessionCookies(taxYearEOY)))
       }
 
-
-      "render the page" when {
-        s"render the page for isAgent: ${scenarioData.isAgent} isWelsh: ${scenarioData.isWelsh} content" which {
-          lazy val result = {
-            dropEmploymentDB()
-            authoriseAgentOrIndividual(scenarioData.isAgent)
-            insertCyaData(EmploymentUserData(
-              sessionId,
-              mtditid,
-              nino,
-              taxYearEOY,
-              employmentId, isPriorSubmission = false, hasPriorBenefits = false, hasPriorStudentLoans = false,
-              EmploymentCYAModel(
-                EmploymentDetails(
-                  employerName = "Whiterun Guards",
-                  employerRef = Some("223/AB12399"),
-                  startDate = Some(startDate),
-                  didYouLeaveQuestion = Some(false),
-                  taxablePayToDate = Some(3000.00),
-                  totalTaxToDate = Some(300.00),
-                  currentDataIsHmrcHeld = false
-                ),
-              )
-            ))
-
-            urlGet(url(taxYearEOY), scenarioData.isWelsh, headers = Seq(HeaderNames.COOKIE -> playSessionCookies(taxYearEOY)))
-          }
-
-          implicit val document: () => Document = () => Jsoup.parse(result.body)
-
-          titleCheck(title, scenarioData.isWelsh)
-        }
-      }
+      result.status shouldBe OK
     }
 
+    "render the page" in {
+      lazy val result = {
+        dropEmploymentDB()
+        authoriseAgentOrIndividual(false)
+        insertCyaData(EmploymentUserData(
+          sessionId,
+          mtditid,
+          nino,
+          taxYearEOY,
+          employmentId, isPriorSubmission = false, hasPriorBenefits = false, hasPriorStudentLoans = false,
+          EmploymentCYAModel(
+            EmploymentDetails(
+              employerName = "Whiterun Guards",
+              employerRef = Some("223/AB12399"),
+              startDate = Some(startDate),
+              didYouLeaveQuestion = Some(false),
+              taxablePayToDate = Some(3000.00),
+              totalTaxToDate = Some(300.00),
+              currentDataIsHmrcHeld = false
+            ),
+          )
+        ))
+
+        urlGet(url(taxYearEOY), false, headers = Seq(HeaderNames.COOKIE -> playSessionCookies(taxYearEOY)))
+      }
+
+      result.status shouldBe OK
+    }
   }
 
   ".submit" should {
@@ -294,8 +152,6 @@ class StudentLoansQuestionControllerISpec extends IntegrationTest with ViewHelpe
 
       "The user is taken to the overview page when the student loans feature switch is off" in {
         val headers = Seq(HeaderNames.COOKIE -> playSessionCookies(taxYearEOY), "Csrf-Token" -> "nocheck")
-
-
         val request = FakeRequest("POST", studentLoansQuestionPage(taxYearEOY, employmentId)).withHeaders(headers: _*)
 
         val result: Future[Result] = {
@@ -322,9 +178,7 @@ class StudentLoansQuestionControllerISpec extends IntegrationTest with ViewHelpe
 
         result.status shouldBe SEE_OTHER
         result.header(HeaderNames.LOCATION) shouldBe Some(appConfig.incomeTaxSubmissionOverviewUrl(taxYearEOY))
-
       }
-
     }
 
     "redirect and update cya data correctly" when {
@@ -365,8 +219,8 @@ class StudentLoansQuestionControllerISpec extends IntegrationTest with ViewHelpe
         lazy val cyaModel = findCyaData(taxYearEOY, employmentId, anAuthorisationRequest).get
         cyaModel.employment.studentLoans.get shouldBe
           StudentLoansCYAModel(uglDeduction = true, Some(1000.22), pglDeduction = false, None)
-
       }
+
       "there is cya data in session without StudentLoans (ugl)" in {
         lazy val form = Map(s"${StudentLoanQuestionForm.studentLoans}[]" -> Seq("ugl"))
         lazy val result = {
@@ -401,7 +255,6 @@ class StudentLoansQuestionControllerISpec extends IntegrationTest with ViewHelpe
         lazy val cyaModel = findCyaData(taxYearEOY, employmentId, anAuthorisationRequest).get
         cyaModel.employment.studentLoans.get shouldBe
           StudentLoansCYAModel(uglDeduction = true, None, pglDeduction = false, None)
-
       }
 
       "there is cya data in session without StudentLoans (pgl)" in {
@@ -437,7 +290,6 @@ class StudentLoansQuestionControllerISpec extends IntegrationTest with ViewHelpe
         lazy val cyaModel = findCyaData(taxYearEOY, employmentId, anAuthorisationRequest).get
         cyaModel.employment.studentLoans.get shouldBe
           StudentLoansCYAModel(uglDeduction = false, None, pglDeduction = true, None)
-
       }
 
       "there is cya data in session without StudentLoans (none)" in {
@@ -474,8 +326,8 @@ class StudentLoansQuestionControllerISpec extends IntegrationTest with ViewHelpe
         lazy val cyaModel = findCyaData(taxYearEOY, employmentId, anAuthorisationRequest).get
         cyaModel.employment.studentLoans.get shouldBe
           StudentLoansCYAModel(uglDeduction = false, None, pglDeduction = false, None)
-
       }
+
       "there is cya data in session without StudentLoans (ugl, pgl)" in {
         lazy val form = Map(s"${StudentLoanQuestionForm.studentLoans}[]" -> Seq("ugl", "pgl"))
         lazy val result = {
@@ -501,7 +353,6 @@ class StudentLoansQuestionControllerISpec extends IntegrationTest with ViewHelpe
           ))
           userDataStub(IncomeTaxUserData(), nino, taxYearEOY)
 
-
           urlPost(url(taxYearEOY), form, follow = false, headers = Seq(HeaderNames.COOKIE -> playSessionCookies(taxYearEOY)))
         }
 
@@ -511,89 +362,69 @@ class StudentLoansQuestionControllerISpec extends IntegrationTest with ViewHelpe
         lazy val cyaModel = findCyaData(taxYearEOY, employmentId, anAuthorisationRequest).get
         cyaModel.employment.studentLoans.get shouldBe
           StudentLoansCYAModel(uglDeduction = true, None, pglDeduction = true, None)
-
       }
     }
 
-    userScenarios.foreach { scenarioData =>
-      import scenarioData.commonExpectedResults._
-      "return form errors" when {
-        s"no option has been selected for isAgent: ${scenarioData.isAgent} isWelsh: ${scenarioData.isWelsh}" which {
-          lazy val form = Map(s"${StudentLoanQuestionForm.studentLoans}[]" -> Seq())
-          lazy val result = {
-            dropEmploymentDB()
-            authoriseAgentOrIndividual(scenarioData.isAgent)
-            insertCyaData(EmploymentUserData(
-              sessionId,
-              mtditid,
-              nino,
-              taxYearEOY,
-              employmentId, isPriorSubmission = false, hasPriorBenefits = false, hasPriorStudentLoans = false,
-              EmploymentCYAModel(
-                EmploymentDetails(
-                  employerName = "Whiterun Guards",
-                  employerRef = Some("223/AB12399"),
-                  startDate = Some(startDate),
-                  didYouLeaveQuestion = Some(false),
-                  taxablePayToDate = Some(3000.00),
-                  totalTaxToDate = Some(300.00),
-                  currentDataIsHmrcHeld = false
-                )
+    "return bad request" when {
+      s"no option has been selected" in {
+        lazy val form = Map(s"${StudentLoanQuestionForm.studentLoans}[]" -> Seq())
+        lazy val result = {
+          dropEmploymentDB()
+          authoriseAgentOrIndividual(false)
+          insertCyaData(EmploymentUserData(
+            sessionId,
+            mtditid,
+            nino,
+            taxYearEOY,
+            employmentId, isPriorSubmission = false, hasPriorBenefits = false, hasPriorStudentLoans = false,
+            EmploymentCYAModel(
+              EmploymentDetails(
+                employerName = "Whiterun Guards",
+                employerRef = Some("223/AB12399"),
+                startDate = Some(startDate),
+                didYouLeaveQuestion = Some(false),
+                taxablePayToDate = Some(3000.00),
+                totalTaxToDate = Some(300.00),
+                currentDataIsHmrcHeld = false
               )
-            ))
-            userDataStub(IncomeTaxUserData(), nino, taxYearEOY)
-            urlPost(url(taxYearEOY), form, welsh = scenarioData.isWelsh, headers = Seq(HeaderNames.COOKIE -> playSessionCookies(taxYearEOY)))
-          }
-
-          "returns the correct status" in {
-            result.status shouldBe BAD_REQUEST
-          }
-
-          implicit val document: () => Document = () => Jsoup.parse(result.body)
-
-          titleCheck(errorPrefix(scenarioData.isWelsh) + title, scenarioData.isWelsh)
-          errorSummaryCheck(errorEmpty, Selectors.checkboxHref)
-          errorAboveElementCheck(errorEmpty)
+            )
+          ))
+          userDataStub(IncomeTaxUserData(), nino, taxYearEOY)
+          urlPost(url(taxYearEOY), form, headers = Seq(HeaderNames.COOKIE -> playSessionCookies(taxYearEOY)))
         }
 
-        s"all options have been selected for isAgent: ${scenarioData.isAgent} isWelsh: ${scenarioData.isWelsh}" which {
-          lazy val form = Map(s"${StudentLoanQuestionForm.studentLoans}[]" -> Seq("ugl", "pgl", "none"))
-          lazy val result = {
-            dropEmploymentDB()
-            authoriseAgentOrIndividual(scenarioData.isAgent)
-            insertCyaData(EmploymentUserData(
-              sessionId,
-              mtditid,
-              nino,
-              taxYearEOY,
-              employmentId, isPriorSubmission = false, hasPriorBenefits = false, hasPriorStudentLoans = false,
-              EmploymentCYAModel(
-                EmploymentDetails(
-                  employerName = "Whiterun Guards",
-                  employerRef = Some("223/AB12399"),
-                  startDate = Some(startDate),
-                  didYouLeaveQuestion = Some(false),
-                  taxablePayToDate = Some(3000.00),
-                  totalTaxToDate = Some(300.00),
-                  currentDataIsHmrcHeld = false
-                )
+        result.status shouldBe BAD_REQUEST
+      }
+
+      s"all options have been selected" in {
+        lazy val form = Map(s"${StudentLoanQuestionForm.studentLoans}[]" -> Seq("ugl", "pgl", "none"))
+        lazy val result = {
+          dropEmploymentDB()
+          authoriseAgentOrIndividual(false)
+          insertCyaData(EmploymentUserData(
+            sessionId,
+            mtditid,
+            nino,
+            taxYearEOY,
+            employmentId, isPriorSubmission = false, hasPriorBenefits = false, hasPriorStudentLoans = false,
+            EmploymentCYAModel(
+              EmploymentDetails(
+                employerName = "Whiterun Guards",
+                employerRef = Some("223/AB12399"),
+                startDate = Some(startDate),
+                didYouLeaveQuestion = Some(false),
+                taxablePayToDate = Some(3000.00),
+                totalTaxToDate = Some(300.00),
+                currentDataIsHmrcHeld = false
               )
-            ))
-            userDataStub(IncomeTaxUserData(), nino, taxYearEOY)
+            )
+          ))
+          userDataStub(IncomeTaxUserData(), nino, taxYearEOY)
 
-            urlPost(url(taxYearEOY), form, welsh = scenarioData.isWelsh, headers = Seq(HeaderNames.COOKIE -> playSessionCookies(taxYearEOY)))
-          }
-
-          "returns the correct status" in {
-            result.status shouldBe BAD_REQUEST
-          }
-
-          implicit val document: () => Document = () => Jsoup.parse(result.body)
-
-          titleCheck(errorPrefix(scenarioData.isWelsh) + title, scenarioData.isWelsh)
-          errorSummaryCheck(errorEmpty, Selectors.checkboxHref)
-          errorAboveElementCheck(errorEmpty)
+          urlPost(url(taxYearEOY), form, headers = Seq(HeaderNames.COOKIE -> playSessionCookies(taxYearEOY)))
         }
+
+        result.status shouldBe BAD_REQUEST
       }
     }
   }
