+++ test/support/ViewHelper.scala
@@ -182,6 +182,18 @@ trait ViewHelper {
     }
   }
 
+  def checkBoxCheck(name: String, selector: String, value: String, checked: Boolean)(implicit document: Document): Unit = {
+    s"'$selector' has a name of '$name'" in {
+      document.select(selector).attr("name") shouldBe name
+    }
+    s"'$selector' has a value of '$value'" in {
+      document.select(selector).attr("value") shouldBe value
+    }
+    s"'$value' has the checked value set to $checked" in {
+      document.select(selector).hasAttr("checked") shouldBe checked
+    }
+  }
+
   def errorSummaryCheck(text: String, href: String)(implicit document: Document): Unit = {
     "contains an error summary" in {
       elementExist(".govuk-error-summary")
